provider_daft/requirements.txt
uvicorn
fastapi
beautifulsoup4
python-dateutil
nums-from-string
ddtrace
https://github.com/AnthonyBloomer/daftlistings/archive/master.zip